CI: Add emulated cross builds.
To catch portability bugs, it is useful to test on architectures with
different endianness and word size.

This commit uses QEMU usermode emulation to test on S390x (64-bit,
BE), ARM (32-bit, LE), and PPC (32-bit, BE).

So now, in all, CI will test on:

* amd64 (64-bit, LE)
* aarch64 (64-bit, LE)
* S390x (64-bit, BE)
* ARM (32-bit, LE)
* PPC (32-bit, BE)

Obviously, all of the non-x86 architectures only build on Linux.