BUILD[NIX](FIX): Move Meson native files out of submodule.
Nix really does not like building flakes with submodules, for some
reason, and the ?submodule=1 hack spontaneously stopped working in
Github actions.

This patch just gives in and copies the files into the repository
proper.

It looks like nix-installer-action@v9 fails under act, so a note to that
effect is added to the workflow file.